+++ CHANGES.md
@@ -18,6 +18,7 @@ of that change.
 - `SequentialEx`, `MergeLayer`, and `res_block` to more easily create resnet and densenet architectures
 - `no_split` method in the data block API
 - `sigmoid_range` function to scale sigmoid to given range, along with `SigmoidRange` layer
+- `DataBunch` performs a sanity check after its initialization and will throw a warning if something is wrong with the data.
 
 ### Changed: